ci: split tests
