+++ tests/functional/spec/US/banner/textColors.test.js
@@ -1,30 +1,10 @@
 import createBannerTest from '../../createBannerTest';
 import accounts from '../accounts';
 
-const positions = ['top', 'left', 'right'];
-const textAlign = ['left', 'right', 'center'];
 const logoTypes = ['primary', 'alternative', 'inline', 'none'];
-const fontSizes = [10, 12, 16];
 const colors = ['black', 'white', 'monochrome', 'grayscale'];
 
 const tests = [].concat(
-    // Logo types with valid logo position options
-    logoTypes
-        .slice(0, 2)
-        .reduce(
-            (array, type) =>
-                array.concat(
-                    positions.map(position => [`${type}, Position:${position}`, { logo: { type, position } }])
-                ),
-            []
-        ),
-    // Logo types that do not have different logo position options
-    logoTypes.slice(2).map(type => [`Logo type:${type}`, { logo: { type } }]),
-    // Each text alignment option
-    textAlign.map(align => [`Text align:${align}`, { text: { align } }]),
-    // Each font size option logo.type-alternative
-    fontSizes.map(size => [`Font size:${size}`, { logo: { type: 'alternative' }, text: { size } }]),
-    // Each logo type, with non-black color options
     colors
         .slice(1)
         .reduce(
@@ -33,13 +13,7 @@ const tests = [].concat(
                     logoTypes.map(type => [`${color} text, Logo type:${type}`, { logo: { type }, text: { color } }])
                 ),
             []
-        ),
-    // Small viewport
-    textAlign.map(align => [
-        'Small viewport',
-        { logo: { type: 'primary' }, text: { position: 'left', align } },
-        { width: 200, height: 100 }
-    ])
+        )
 );
 
 // +1 is for GPL unqualified
